+++ packages/canner-helpers/src/index.js
@@ -2,6 +2,7 @@
 export Item from './item';
 export Context from './context';
 export BackButton from './BackButton';
+export Buttons from './Buttons';
 export ConfirmButton from './confirmButton';
 export ResetButton from './cancelButton';
 export LiteCMS from './liteCMS';
